pyproject.toml
[project]
name = "def-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "afip-py>=1.2.0",
    "asyncpg>=0.30.0",
    "fastapi[standard]>=0.119.0",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.11.0",
    "python-jose[cryptography]>=3.5.0",
]
